+++ packages/sdk/akamai-edgekv/src/index.ts
@@ -12,23 +12,20 @@
 import {
   init as initEdge,
   LDClient,
-  LDContext,
-  LDMultiKindContext,
-  LDSingleKindContext,
   LDOptions,
   EdgeFeatureStore,
 } from '@launchdarkly/akamai-edgeworker-sdk-common';
 import { BasicLogger } from '@launchdarkly/js-server-sdk-common';
 import EdgeKVProvider from './edgekv/edgeKVProvider';
 
-type AkamaiLDClientParams = {
+export * from '@launchdarkly/akamai-edgeworker-sdk-common';
+
+export type AkamaiLDClientParams = {
   sdkKey: string;
   options?: LDOptions;
   namespace: string;
   group: string;
 };
-
-export type { LDClient, LDContext, LDMultiKindContext, LDSingleKindContext };
 
 /**
  * Initialize Launchdarkly client using Akamai's Edge KV as a feature store
@@ -51,6 +48,6 @@ export const init = ({
     edgeFeatureStore: featureStore,
     platformName: 'Akamai EdgeWorker',
     sdkName: '@launchdarkly/akamai-server-edgekv-sdk',
-    sdkVersion: '0.1.0', // {x-release-please-version}
+    sdkVersion: '0.3.1', // {x-release-please-version}
   });
 };