+++ packages/sdk/vercel/src/createPlatformInfo.ts
@@ -10,7 +10,7 @@ class VercelPlatformInfo implements Info {
   sdkData(): SdkData {
     return {
       name: '@launchdarkly/vercel-server-sdk',
-      version: '0.4.0', // {x-release-please-version}
+      version: '0.4.3', // {x-release-please-version}
     };
   }
 }