+++ packages/shared/sdk-server/src/store/index.ts
@@ -1,4 +1,5 @@
 import AsyncStoreFacade from './AsyncStoreFacade';
+import PersistentDataStoreWrapper from './PersistentDataStoreWrapper';
 import { deserializePoll } from './serialization';
 
-export { AsyncStoreFacade, deserializePoll };
+export { AsyncStoreFacade, PersistentDataStoreWrapper, deserializePoll };