+++ packages/sdk/akamai-edgekv/CHANGELOG.md
@@ -2,6 +2,50 @@
 
 All notable changes to the LaunchDarkly SDK for Akamai Workers will be documented in this file. This project adheres to [Semantic Versioning](https://semver.org).
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^0.1.0 to ^0.1.1
+    * @launchdarkly/js-server-sdk-common bumped from ^1.0.2 to ^1.0.3
+
+## [0.3.1](https://github.com/launchdarkly/js-core/compare/akamai-server-edgekv-sdk-v0.3.0...akamai-server-edgekv-sdk-v0.3.1) (2023-06-13)
+
+
+### Bug Fixes
+
+* Update readme links for Akamai SDKs ([#157](https://github.com/launchdarkly/js-core/issues/157)) ([8a138c1](https://github.com/launchdarkly/js-core/commit/8a138c169cc33c1e155c1d3643a73ce55d16b902))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^0.2.0 to ^0.2.1
+    * @launchdarkly/js-server-sdk-common bumped from ^1.0.3 to ^1.0.4
+
+## [0.3.0](https://github.com/launchdarkly/js-core/compare/akamai-server-edgekv-sdk-v0.2.1...akamai-server-edgekv-sdk-v0.3.0) (2023-06-09)
+
+
+### Features
+
+* Release beta version of Akamai SDKs ([#155](https://github.com/launchdarkly/js-core/issues/155)) ([aa8158c](https://github.com/launchdarkly/js-core/commit/aa8158cbf16a9abf681438341b52dc1f640c5c40))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/akamai-edgeworker-sdk-common bumped from ^0.1.1 to ^0.2.0
+
+## [0.2.0](https://github.com/launchdarkly/js-core/compare/akamai-server-edgekv-sdk-v0.1.0...akamai-server-edgekv-sdk-v0.2.0) (2023-06-06)
+
+
+### Features
+
+* Exported typescript types for Akamai SDK ([e7be20c](https://github.com/launchdarkly/js-core/commit/e7be20caa2f13abecf12a29ddc46a45a59fc7b5c))
+* Exported typescript types for Akamai SDK ([#135](https://github.com/launchdarkly/js-core/issues/135)) ([01da2a9](https://github.com/launchdarkly/js-core/commit/01da2a9abd203ed171497c92eab8abd0f9aa81c2))
+
 ## 0.1.0 (2023-06-02)