+++ packages/shared/akamai-edgeworker-sdk/CHANGELOG.md
@@ -1,3 +1,29 @@
 # Changelog
 
 All notable changes to the LaunchDarkly SDK for Akamai Workers will be documented in this file. This project adheres to [Semantic Versioning](https://semver.org).
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common bumped from ^1.0.2 to ^1.0.3
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @launchdarkly/js-server-sdk-common bumped from ^1.0.3 to ^1.0.4
+
+## [0.2.0](https://github.com/launchdarkly/js-core/compare/akamai-edgeworker-sdk-common-v0.1.1...akamai-edgeworker-sdk-common-v0.2.0) (2023-06-09)
+
+
+### Features
+
+* Release beta version of Akamai SDKs ([#155](https://github.com/launchdarkly/js-core/issues/155)) ([aa8158c](https://github.com/launchdarkly/js-core/commit/aa8158cbf16a9abf681438341b52dc1f640c5c40))
+
+## 0.1.0 (2023-06-06)
+
+
+### Features
+
+* Release akamai-edgeworker-sdk-common ([#138](https://github.com/launchdarkly/js-core/issues/138)) ([f22a680](https://github.com/launchdarkly/js-core/commit/f22a680be3eb8bb2af543ff1d5bc5198bb8a6a05))